requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pytz==2023.3
httpx==0.25.0
aiohttp==3.9.1
python-multipart==0.0.6
requests==2.31.0

